--- packages/brightcove-video/__tests__/android/brightcove-video.android.test.js
@@ -1,7 +1,0 @@
-import shared from "../shared";
-import sharedNative from "../shared-native";
-
-describe("BrightcoveVideo on android", () => {
-  shared();
-  sharedNative();
-});
