--- packages/brightcove-video/__tests__/android/jest.config.js
@@ -1,3 +1,0 @@
-const jestConfigurator = require("@times-components/jest-configurator").default;
-
-module.exports = jestConfigurator("android", __dirname);
