+++ packages/brightcove-video/__tests__/shared.js
@@ -2,8 +2,9 @@ import "react-native";
 import React from "react";
 import { shallow } from "enzyme";
 import BrightcoveVideo from "../src/brightcove-video";
+import PlayIcon from "../src/play-icon";
 
-export const defaultProps = {
+export const defaultVideoProps = {
   accountId: "[account id]",
   playerId: "[player id]",
   videoId: "[video id]",
@@ -17,7 +18,17 @@ export const defaultProps = {
 
 export default () => {
   it("renders correctly", () => {
-    const tree = shallow(<BrightcoveVideo {...defaultProps} />);
+    const tree = shallow(<BrightcoveVideo {...defaultVideoProps} />);
     expect(tree).toMatchSnapshot();
-  })
+  });
+
+  it("renders correctly without a poster image", () => {
+    const tree = shallow(<BrightcoveVideo {...defaultVideoProps} poster={null} />);
+    expect(tree).toMatchSnapshot();
+  });
+
+  it("renders a play icon correctly", () => {
+    const tree = shallow(<PlayIcon />);
+    expect(tree).toMatchSnapshot();
+  });
 };
