+++ packages/brightcove-video/old__tests__/brightcove-video.shared.js
@@ -47,36 +47,6 @@ const brightcoveSharedTests = BrightcoveVideo => {
     expect(root.toJSON()).toMatchSnapshot();
   });
 
-  const testSubscriberAndVideoPaidStatus = (
-    subscriberIsPaid,
-    videoIsPaidOnly
-  ) => {
-    const tree = renderer
-      .create(
-        <IsPaidSubscriber.Provider value={subscriberIsPaid}>
-          <BrightcoveVideo {...defaultVideoProps} paidonly={videoIsPaidOnly} />
-        </IsPaidSubscriber.Provider>
-      )
-      .toJSON();
-    expect(tree).toMatchSnapshot();
-  };
-
-  it("renders a paidonly video correctly for unpaid users", () => {
-    testSubscriberAndVideoPaidStatus(false, true);
-  });
-
-  it("renders a paidonly video correctly for paid users", () => {
-    testSubscriberAndVideoPaidStatus(true, true);
-  });
-
-  it("renders a non-paidonly video correctly for unpaid users", () => {
-    testSubscriberAndVideoPaidStatus(false, false);
-  });
-
-  it("renders a non-paidonly video correctly for paid users", () => {
-    testSubscriberAndVideoPaidStatus(true, false);
-  });
-
   it("pauses other playing videos if play is called", () => {
     renderer.create(<BrightcoveVideo {...defaultVideoProps} />);
     renderer.create(<BrightcoveVideo {...defaultVideoProps} />);
